Implement constant and function application expression parser

+++ include/pebble/syntax/expression.hpp
@@ -25,14 +25,14 @@ namespace pebble {
       class float_const_expr
       {
       private:
-        float data_;
+        double data_;
 
       public:
         float_const_expr() = default;
-        explicit float_const_expr(float d): data_(d) {}
+        explicit float_const_expr(double d): data_(d) {}
 
         std::string to_string() const { return std::to_string(data_); }
-        float data() const { return data_; }
+        double data() const { return data_; }
       };
 
       class ident_expr
